:hammer: Move convertToCSV tests from exporters to statements
